requirements-docs.txt
# minimum needed to build jupyterlite docs... in addition to nodejs and yarn
# see .binder/ and docs/ for full development/docs environments
-r requirements-test.txt

# build
docutils >=0.18
jupyter-server-mathjax
myst-nb
pydata-sphinx-theme
sphinx >=5.1,<6
sphinx-autodoc-typehints
sphinx-jsonschema
sphinxext-rediraffe

# check
pytest-check-links
